src/app-Qdrant-Visualizer/requirements.txt
qdrant-client
numpy
scipy
umap-learn
scikit-learn
dash
fastapi
uvicorn
plotly
pandas
